Replace spec/dummy app with a simpler fake Rails app
In prior Rails versions, it used to be more challenging to create a
stand-alone application without the entire file structure of Rails
surrounding it, which is why the spec/dummy was in here in the first
place.

This is no longer the case, and this change removes the dummy app in
favor of a smaller set of files which achieve the same result of
providing a small footprint Rails app which the test suite can load the
gem into, and run tests against an actual Rails app.
