+++ .gitignore
@@ -1,9 +1,6 @@
 .bundle/
 log/*.log
 pkg/
-spec/dummy/db/*.sqlite3
-spec/dummy/log/*.log
-spec/dummy/tmp/
 *.swp
 *.gem
 gemfiles/*.lock
