--- spec/dummy/app/controllers/application_controller.rb
@@ -1,3 +1,0 @@
-class ApplicationController < ActionController::Base
-  protect_from_forgery
-end
