--- spec/dummy/config.ru
@@ -1,4 +1,0 @@
-# This file is used by Rack-based servers to start the application.
-
-require ::File.expand_path('../config/environment',  __FILE__)
-run Dummy::Application
