--- spec/dummy/config/boot.rb
@@ -1,10 +1,0 @@
-require 'rubygems'
-gemfile = File.expand_path('../../../../Gemfile', __FILE__)
-
-if File.exist?(gemfile)
-  ENV['BUNDLE_GEMFILE'] = gemfile
-  require 'bundler'
-  Bundler.setup
-end
-
-$:.unshift File.expand_path('../../../../lib', __FILE__)
