--- spec/dummy/config/environment.rb
@@ -1,5 +1,0 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
-
-# Initialize the rails application
-Dummy::Application.initialize!
