--- spec/dummy/public/javascripts/application.js
@@ -1,2 +1,0 @@
-// Place your application-specific JavaScript functions and classes here
-// This file is automatically included by javascript_include_tag :defaults
