+++ spec/fixtures/app/views/layouts/alternate.html.erb
