+++ spec/fixtures/app/views/layouts/application.html.erb
