+++ spec/fixtures/app/views/pages/also_dir/also_nested.html.erb
