+++ spec/fixtures/app/views/pages/exists_but_references_nonexistent_partial.html.erb
