+++ spec/fixtures/config/database.yml
@@ -1,0 +1,3 @@
+test:
+  adapter: sqlite3
+  database: ":memory:"
