+++ spec/support/app/controllers/subclassed_pages_controller.rb
